+++ src/c/profile/session/create_entities_ref.c
@@ -29,7 +29,7 @@ inline uint16_t create_entity_ref(mrSession* session, mrStreamId stream_id,
                                   CREATE_Payload* payload)
 {
     uint16_t returned_value = MR_INVALID_REQUEST_ID;
-    uint32_t ref_size = strlen(ref) + 1;
+    uint32_t ref_size = (uint32_t)strlen(ref) + 1;
     if(ref_size <= UINT16_MAX)
     {
         // Use participant access to access to the ref base of any object variant. //Future elegant change?
